+++ helper.js
@@ -1,3 +1,5 @@
+const request = require('request-promise-native');
+
 exports.extractEmailAddress = function(slackFieldValue) {
   var regex = /<mailto:([^|]*?)(?:\|.*)?>/i;
   var matchResults = slackFieldValue.trim().match(regex);
@@ -22,3 +24,37 @@ exports.splitFullName = function(fullName) {
     return ['', ''];
   }
 };
+
+exports.badRequestResponse =  function(message, cors) {
+  const log = `Bad request: ${message}`;
+  console.log(log);
+
+  const result = {
+    statusCode: 400,
+    body: log
+  };
+
+  if (cors) {
+    result.headers = {
+      'Access-Control-Allow-Origin': '*'
+    };
+  }
+
+  return result;
+}
+
+exports.errorToSlack = async function(responseUrl, err) {
+  try {
+    return await request.post({
+      url: responseUrl,
+      body: {
+        'response_type': 'ephemeral',
+        'replace_original': false,
+        'text': 'Error: ' + err.toString()
+      },
+      json: true
+    });
+  } catch (err) {
+    console.error('Failed to send error to Slack: ', err);
+  }
+}
