+++ serverless.yml
@@ -65,7 +65,7 @@ resources:
 
 functions:
   createLetter:
-    handler: handler.createLetter
+    handler: createLetter.entryPoint
     events:
       - http:
           path: submit
@@ -74,7 +74,7 @@ functions:
           cors: true
 
   approveLetter:
-    handler: handler.approveLetter
+    handler: approveLetter.entryPoint
     events:
       - http:
           path: approve
@@ -82,7 +82,7 @@ functions:
           private: false
 
   slash:
-    handler: handler.slash
+    handler: slash.invokeSlashCommand
     events:
       - http:
           path: slash
@@ -92,7 +92,7 @@ functions:
       SLASH_TOPIC: ${self:custom.ahvSlashTaskQueue}
 
   processSlashCommand:
-    handler: handler.processSlashCommand
+    handler: slash.processSlashCommand
     events:
       - sns: ahvSlashTasks
     timeout: 180
